Update ppmdata.R
Hi Skip, I think that there are lots of simplifications that could be made for quasirandomMethod.  Much of what you have is designed for BAS, a task that we don't have to consider here.  The code I sent (via email) will be more direct, but will also induce a dependency on MBHdesign.

Gotta run now, sorry.